moves program form to client/components/pages/submission-system and split up mutation gql
splits up validation schema
