+++ client/components/pages/submission-system/create-program/CREATE_PROGRAM_MUTATION.gql
@@ -12,7 +12,3 @@ mutation CREATE_PROGRAM_MUTATION($program: ProgramInput!) {
     membershipType
   }
 }
-
-mutation UPDATE_PROGRAM_MUTATION($updates: UpdateProgramInput!, $shortName: String!) {
-  updateProgram(shortName: $shortName, updates: $updates)
-}
