+++ client/components/pages/submission-system/program-management/UPDATE_PROGRAM_MUTATION.gql
@@ -1,0 +1,3 @@
+mutation UPDATE_PROGRAM_MUTATION($updates: UpdateProgramInput!, $shortName: String!) {
+  updateProgram(shortName: $shortName, updates: $updates)
+}
